Move CpgLoaderConfig to its own file
